Move step layout to own folder
[MAILPOET-2142]
