+++ assets/js/src/wizard/layout/step_layout.jsx
